pyproject.toml
[tool.poetry]
name = "nunet"
version = "0.1.0"
description = ""
authors = ["Elder Millenial <eldermillenial@protonmail.com>"]
readme = "README.md"
packages = [{ include = "nunet", from = "src" }]

[tool.poetry.dependencies]
python = "^3.10"
websockets = "^11.0.3"
requests = "^2.31.0"
python-dotenv = "^1.0.0"
pycardano = { git = "https://github.com/theeldermillenial/pycardano.git" }
types-requests = "^2.31.0.10"
pydantic = "^2.4.2"


[tool.poetry.group.dev.dependencies]
pytest = "^7.4.2"
mypy = "^1.6.0"
ruff = "^0.0.292"
pre-commit = "^3.5.0"
flake8 = "^6.1.0"
black = "^23.9.1"
bump2version = "^1.0.1"
isort = "^5.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"
